refactor: move buyer and seller functions into policy class
